+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: seongjch <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/25 15:27:15 by seongjch          #+#    #+#             */
-/*   Updated: 2022/04/09 18:14:57 by seongjch         ###   ########.fr       */
+/*   Updated: 2022/04/09 18:26:04 by seongjch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,6 +23,6 @@ int		ft_putstr_fd(char *s, int fd);
 int		ft_putnbr_fd(long long n, int fd);
 int		ft_puthexlow_fd(unsigned int ln, int fd);
 int		ft_puthexup_fd(unsigned int ln, int fd);
-int		ft_putpoint_fd(unsigned int ln, int fd);
+int		ft_putpoint_fd(unsigned long long ln, int fd);
 
 #endif
