+++ ft_putpoint_fd.c
@@ -6,13 +6,13 @@
 /*   By: seongjch <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/25 19:51:24 by seongjch          #+#    #+#             */
-/*   Updated: 2022/04/09 18:16:49 by seongjch         ###   ########.fr       */
+/*   Updated: 2022/04/09 18:29:42 by seongjch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_putpoint_fd(unsigned int ln, int fd)
+int	ft_putpoint_fd(unsigned long long ln, int fd)
 {
 	int	cnt;
